# open-api/passtheo-content-service-openapi.yaml
openapi: 3.1.0
info:
  title: PassTheo Content Service
  description: |
    The core learning engine for PassTheo. Serves questions, tracks progress,
    runs mock exams, manages streaks, awards achievements, and generates study plans.
    
    Content hierarchy: Country → ProductType → Product → Domain → Topic → Question
    Language is Strapi i18n, orthogonal to hierarchy.
  version: 1.0.0
servers:
  - url: http://localhost:8087
    description: Local development
...
paths:
...
  /api/achievements/me:
    get:
      operationId: getMyAchievements
      tags: [Achievements]
      summary: Get earned + locked achievements
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
        - name: productCode
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LocaleQuery'
      responses:
        '200':
          description: Achievement gallery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AchievementList'

  # ─── STUDY PLAN ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
    UserHeader:
      name: X-Keycloak-User-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    LocaleQuery:
      name: locale
      in: query
      schema:
        type: string
        default: "nl"
      description: Content locale (nl, en, ar, tr, pl)
  schemas:
...
    ApiResponse_AchievementList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/AchievementDto' } }
        timestamp: { type: string, format: date-time }
...
    AchievementDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        isEarned: { type: boolean }
        earnedAt: { type: string, format: date-time, nullable: true }
        triggerType: { type: string }
        triggerValue: { type: integer }
        currentProgress: { type: integer, description: "user's current value toward this trigger" }
        progressPercent: { type: number }
        productCode: { type: string, nullable: true, description: "null = platform-wide achievement" }
...